GrowingBitSet implementation, update algorithms

+++ src/main/scala/graph/EdgeUtils.scala
@@ -7,6 +7,6 @@ package graph
  * wedge2edge: convert weighted edge to simple edge
  */
 object EdgeUtils {
-  def edge2wedge(e: SimpleEdge) = Edge(e.from, e.to, util.Random.nextFloat)
+  def edge2wedge(e: SimpleEdge) = Edge(e.from, e.to, util.Random.nextDouble)
   def wedge2edge(e: WeightedEdge) = Edge(e.from, e.to)
 }
