+++ src/main/scala/helper/GrowingArray.scala
@@ -29,7 +29,7 @@ class GrowingArray[T: Manifest](v: T) {
   }
 
   def size = data.size << rowScale
-  def unVisited(index: Int) = this(index) == default
+  def unVisited(index: Long) = this(index) == default
   def nUpdated = data.view.map { _.count { _ != default } }.sum
   def updated = {
     val i = Iterator.from(0)
